creating EmbeddedLegacyEditors.tsx; Toolbar Menu;
